pkg/directory: deprecate, and move to internal
