+++ volume/service/convert.go
@@ -10,7 +10,7 @@ import (
 	"github.com/docker/docker/api/types/filters"
 	volumetypes "github.com/docker/docker/api/types/volume"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/pkg/directory"
+	"github.com/docker/docker/internal/directory"
 	"github.com/docker/docker/volume"
 )
